Refactored ArticleDisplayData -> ArticleDisplayDataParcelable (will need to pull remaining parcelable references out of ArticleListPresenter)
