+++ app/src/main/java/com/chrisfry/nerdnews/userinterface/interfaces/ArticleSelectionListener.kt
@@ -1,6 +1,6 @@
 package com.chrisfry.nerdnews.userinterface.interfaces
 
-import com.chrisfry.nerdnews.model.ArticleDisplayModel
+import com.chrisfry.nerdnews.model.ArticleDisplayModelParcelable
 
 /**
  * Listener for when an article is selected
@@ -11,5 +11,5 @@ interface ArticleSelectionListener {
      *
      * @param article: Article data for the selected article
      */
-    fun onArticleSelected(article: ArticleDisplayModel)
+    fun onArticleSelected(article: ArticleDisplayModelParcelable)
 }
